No need to override the html complete file anymore
The reason I had this file was because the html omni-completion file
added an initial quote on attributes so it became class=" instead of
just class and when I added the closing quote auto-pairs added 2 quotes
so it ended up being class=""". However jiangmiao fixed so that
auto-pairs won't add the second quote if a initial quote already exists.